supress line goes to 0
the feature worked with gnuplot, now also with matplot

teh limits are: do NOT draw line down  (ymax-ymin)*0.3 < value  <= ymin.  same for ymax
eg
if you set yMin to 0.0001 and yMax to 4, line is not drawn if y =0 but would be draw  if < -1.2  ie significantly.
assuming that you want to suppress the line to 0.